+++ platform/ios/src/UIView+MGLAdditions.h
@@ -1,0 +1,19 @@
+#import <UIKit/UIKit.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface UIView (MGLAdditions)
+
+- (UIViewController *)mgl_viewControllerForLayoutGuides;
+
+- (NSLayoutYAxisAnchor *)mgl_safeTopAnchor;
+
+- (NSLayoutXAxisAnchor *)mgl_safeLeadingAnchor;
+
+- (NSLayoutYAxisAnchor *)mgl_safeBottomAnchor;
+
+- (NSLayoutXAxisAnchor *)mgl_safeTrailingAnchor;
+
+@end
+
+NS_ASSUME_NONNULL_END